projection improvements

+++ src/main/kotlin/io/github/abaddon/kcqrs/core/domain/messages/events/IEvent.kt
@@ -3,5 +3,5 @@ package io.github.abaddon.kcqrs.core.domain.messages.events
 import io.github.abaddon.kcqrs.core.domain.messages.IMessage
 
 interface IEvent: IMessage {
-    val version: Int
+    val version: Long
 }
